Ported function from 'refund' package

A modified version of fosr.vs, but with groups parameter to allow grouping time points rather than the whole coefficient when the underlying functions are locally supported.

Usage

fosr_vs(
  formula,
  data,
  nbasis = 10,
  method = c("ls", "grLasso", "grMCP", "grSCAD"),
  epsilon = 1e-05,
  max.iter_num = 100,
  groups = NULL
)

Arguments

formula, data, nbasis, method, epsilon, max.iter_num

see fosr.vs

groups

integer vector with length of number of time-points of how time-points should be grouped; default is NULL, indicating there is no local sparsity.